+++ dependabot/update_dependencies_version.py
@@ -2,12 +2,11 @@ import json
 import os
 import sys
 import time
-import urllib.request
 
 from github import Github, InputGitAuthor, GithubException
-from retry import retry
 
 import constants
+import utils
 
 LANG_VERSION_KEY = 'ballerinaLangVersion'
 LANG_VERSION_UPDATE_BRANCH = 'automated/dependency_version_update'
@@ -66,7 +65,11 @@ def main():
     global extensions_file
     global all_modules
 
-    extensions_file = get_extensions_file()
+    try:
+        extensions_file = utils.get_extensions_file()
+    except Exception as e:
+        print('[Error] Error while loading modules list ', e)
+        sys.exit(1)
 
     print("Workflow invoked of type '" + event_type + "'")
     if event_type == 'schedule' and not extensions_file['auto_bump']:
@@ -86,48 +89,7 @@ def get_lang_version():
     if override_ballerina_version != '':
         return override_ballerina_version
     else:
-        try:
-            version_string = open_url(
-                'https://api.github.com/orgs/ballerina-platform/packages/maven/org.ballerinalang.jballerina-tools/versions').read()
-        except Exception as e:
-            print('[Error] Failed to get ballerina packages version', e)
-            sys.exit(1)
-
-        versions_list = json.loads(version_string)
-        latest_version = versions_list[0]['name']
-        if extensions_file['lang_version_regex'] != "":
-            for version in versions_list:
-                version_name = version['name']
-                if extensions_file['lang_version_regex'] in version_name:
-                    latest_version = version_name
-                    break
-        return latest_version
-
-
-@retry(
-    urllib.error.URLError,
-    tries=constants.HTTP_REQUEST_RETRIES,
-    delay=constants.HTTP_REQUEST_DELAY_IN_SECONDS,
-    backoff=constants.HTTP_REQUEST_DELAY_MULTIPLIER
-)
-def open_url(url):
-    request = urllib.request.Request(url)
-    request.add_header('Accept', 'application/vnd.github.v3+json')
-    request.add_header('Authorization', 'Bearer ' + ballerina_bot_token)
-
-    return urllib.request.urlopen(request)
-
-
-def get_extensions_file():
-    try:
-        with open(constants.EXTENSIONS_FILE) as f:
-            module_list = json.load(f)
-
-    except Exception as e:
-        print('[Error] Error while loading modules list ', e)
-        sys.exit(1)
-
-    return module_list
+        return utils.get_latest_lang_version(ballerina_bot_token)
 
 
 def check_and_update_lang_version():
@@ -343,9 +305,9 @@ def check_pending_build_checks(index: int):
             current_level_modules[index][MODULE_CONCLUSION] = MODULE_CONCLUSION_BUILD_RELEASED
         else:
             try:
-                packages_list_string = open_url(
+                packages_list_string = utils.open_url(
                     'https://api.github.com/orgs/' + constants.BALLERINA_ORG_NAME + '/packages/maven/' + module[
-                        'group_id'] + '.' + module['artifact_id'] + '/versions').read()
+                        'group_id'] + '.' + module['artifact_id'] + '/versions', ballerina_bot_token).read()
                 packages_list = json.loads(packages_list_string)
                 latest_package = packages_list[0]['name']
